+++ demo/Filter/FilterDemo.jsx
@@ -9,58 +9,31 @@
 import React from 'react';
 import Filter from 'salt-filter';
 import List from 'salt-list'
-import Icon from 'salt-icon'
+import Button from 'salt-button'
 import AngleRight from 'salt-icon/lib/AngleRight';
+import Switch from 'salt-switch'
+
+const ButtonGroup = Button.ButtonGroup
 
 class TestView extends React.Component {
   constructor(props) {
     super(props)
     this.state = {
-      listData: [
-        {
-          imgUrl: 'https://img.alicdn.com/tps/TB15YftJFXXXXafXpXXXXXXXXXX-699-698.jpg',
-          text: '前端开发工程师',
-          title: '马慧（穆心）',
-        },
-        {
-          imgUrl: 'https://img.alicdn.com/tps/TB15YftJFXXXXafXpXXXXXXXXXX-699-698.jpg',
-          text: '资深交互设计师',
-          title: '周姮',
-        },
-        {
-          imgUrl: 'https://img.alicdn.com/tps/TB1P1vaLpXXXXXxaXXXXXXXXXXX-50-50.jpg',
-          text: '交互设计师',
-          title: '郝晓敏 (钰馨）',
-          date: '2017-7-8',
-        },
-        {
-          imgUrl: 'https://img.alicdn.com/tps/TB1P1vaLpXXXXXxaXXXXXXXXXXX-50-50.jpg',
-          text: '交互设计师',
-          title: '张三丰（曾金）',
-          date: '2017-7-8',
-        }
-      ]
+      on1: false
     };
   }
 
   onClick = (e) => {
-    this.props.onChange({'aa': [{text: 'xxx', value: 'xxxx'}]})
+    const { on1 } = this.state
+    this.setState({
+      on1: !on1
+    })
+    this.props.onChange([{text: 'xxx', value: 'xxxx'}, {text: 'aaa', value: 'bbbb'}])
   }
 
   render() {
     return (
-      <div>
-        <p style={{ lineHeight: '40px', height: '40px' }}>自定义渲染</p>
-        <List
-          layout="left"
-          hasRightIcon
-          icon={<AngleRight/>}
-          iconWidth={20}
-          isDelete
-          data={this.state.listData}
-          onClick={this.onClick}
-        />
-      </div>
+      <Switch on={this.state.on1} onChange={this.onClick} />
     )
   }
 }
@@ -75,7 +48,7 @@ class Demo extends React.Component {
       defaultFilterValue: {
         sort: [
           {
-            text: '默认排序',
+            text: '默认',
             value: ''
           },
         ],
@@ -115,7 +88,8 @@ class Demo extends React.Component {
           title: '张三丰（曾金）',
           date: '2017-7-8',
         }
-      ]
+      ],
+      tip: '提示'
     };
   }
 
@@ -134,10 +108,43 @@ class Demo extends React.Component {
     console.log(dataItem);
   }
 
+  onClear = () => {
+    this.filter.clearValue()
+    this.setState({
+      tip: 'Filter被清空了'
+    })
+  }
+  onGet = () => {
+    const value = this.filter.getValue()
+    this.setState({
+      tip: 'Filter 内容：' + JSON.stringify(value)
+    })
+  }
+  onSet = () => {
+    this.filter.setValue('brand4', [
+      {
+        text: '阿里巴巴',
+        value: 'alibaba'
+      },
+      {
+        text: '阿里巴巴2',
+        value: 'alibaba2'
+      }
+    ])
+    this.setState({
+      tip: 'Filter已经被设置了！'
+    })
+  }
+
   render() {
     return (
       <div>
+        <div style={{ lineHeight: 2, margin: '10px' }}>
+          <p>{this.state.tip}</p>
+          <p>something else</p>
+        </div>
         <Filter
+          ref={(c) => {this.filter = c}}
           size={4}
           activeIndex={1}
           defaultValue = {this.state.defaultFilterValue}
@@ -145,15 +152,15 @@ class Demo extends React.Component {
             [
               {
                 name: 'sort',
-                title: '默认排序',
+                title: '默认',
                 type: 'order',  // select | order | switch
                 items: [
                   {
-                    text: '默认排序',
+                    text: '默认',
                     value: ''
                   },
                   {
-                    text: '距离',
+                    text: () => '距离',
                     value: 'distance'
                   },
                   {
@@ -189,7 +196,7 @@ class Demo extends React.Component {
                     value: 'grade'
                   },
                   {
-                    text: '50-100人',
+                    text: () => '10-25人',
                     value: 'cainiao'
                   },
                   {
@@ -203,50 +210,8 @@ class Demo extends React.Component {
                 ]
               },
               {
-                name: 'brand',
-                title: '多个选择项',
-                type: 'select',
-                maxLine: 2,
-                multiSelect: true,
-                items: [
-                  {
-                    text: '阿里巴巴',
-                    value: 'alibaba'
-                  },
-                  {
-                    text: '淘宝',
-                    value: 'taobao'
-                  },
-                  {
-                    text: '菜鸟',
-                    value: 'cainiao'
-                  },
-                  {
-                    text: '优酷',
-                    value: 'youku'
-                  },
-                  {
-                    text: '天猫',
-                    value: 'tianmao'
-                  },
-                  {
-                    text: '百度',
-                    value: 'baidu'
-                  },
-                  {
-                    text: '腾讯',
-                    value: 'tengxun'
-                  },
-                  {
-                    text: '有道',
-                    value: 'youdao'
-                  }
-                ]
-              },
-
-              {
-                name: 'brand2',
-                title: '单选',
+                name: 'brand3',
+                title: '单选品牌',
                 type: 'select',
                 maxLine: 2,
                 multiSelect: false,
@@ -286,54 +251,14 @@ class Demo extends React.Component {
                 ]
               },
               {
-                name: 'brand3',
-                title: '品牌',
-                type: 'select',
-                maxLine: 2,
-                multiSelect: true,
-                items: [
-                  {
-                    text: '阿里巴巴',
-                    value: 'alibaba'
-                  },
-                  {
-                    text: '淘宝',
-                    value: 'taobao'
-                  },
-                  {
-                    text: '菜鸟',
-                    value: 'cainiao'
-                  },
-                  {
-                    text: '优酷',
-                    value: 'youku'
-                  },
-                  {
-                    text: '天猫',
-                    value: 'tianmao'
-                  },
-                  {
-                    text: '百度',
-                    value: 'baidu'
-                  },
-                  {
-                    text: '腾讯',
-                    value: 'tengxun'
-                  },
-                  {
-                    text: '有道',
-                    value: 'youdao'
-                  }
-                ]
-              },
-              {
                 name: 'someOther',
+                multiSelect: true,
                 title: '自定义渲染',
                 renderView: TestView,
               },
               {
                 name: 'brand4',
-                title: '品牌',
+                title: '多选品牌',
                 type: 'select',
                 maxLine: 3,
                 multiSelect: true,
@@ -378,7 +303,21 @@ class Demo extends React.Component {
               }
             ]
           }
-          onChange={(data) => {
+          onChange={(data, filter) => {
+            if (data.name ==='quickSort') {
+              console.log(filter.getValue())
+              // filter.clearValue()
+              // filter.setValue('brand4', [
+              //   {
+              //     text: '阿里巴巴',
+              //     value: 'alibaba'
+              //   },
+              //   {
+              //     text: '阿里巴巴2',
+              //     value: 'alibaba2'
+              //   }
+              // ])
+            }
             // can do confirm
             console.log('on change: ', data)
             switch (data.name) {
@@ -393,11 +332,11 @@ class Demo extends React.Component {
               default:
             }
           }}
-          onConfirm={(data) => {
+          onConfirm={(data, filter) => {
             console.log('on confirm: ', data)
             // do confirm
           }}
-          onReset={(data) => {
+          onReset={(data, filter) => {
             console.log('on reset: ', data)
             // do something
           }}
@@ -413,6 +352,11 @@ class Demo extends React.Component {
           onDelete={this.handleDelete}
           data={this.state.listData}
         />
+        <ButtonGroup>
+          <Button type="secondary" display="inline" onClick={this.onClear}>清空</Button>
+          <Button type="secondary" display="inline" onClick={this.onGet}>获取</Button>
+          <Button type="primary" display="inline" onClick={this.onSet}>设置</Button>
+        </ButtonGroup>
       </div>
     );
   }