+++ src/RateField/RateField.jsx
@@ -29,7 +29,6 @@ class RateField extends React.Component {
     return (
       <Field
         {...Field.getFieldProps(t.props)}
-        layout="h"
         className={classnames(Context.prefixClass('rate-field'), {
           [t.props.className]: !!t.props.className,
           readOnly: t.props.readOnly,