+++ src/Textarea/Textarea.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import calculateHeight from './calculateHeight';
 import { shouldUpdate } from '../Utils';
 
@@ -7,6 +8,13 @@ export default class Textarea extends React.Component {
     onChange: () => {},
     minRows: 1,
     maxRows: 10,
+    rows: undefined,
+  }
+  static propTypes = {
+    minRows: PropTypes.number,
+    maxRows: PropTypes.number,
+    onChange: PropTypes.func,
+    rows: PropTypes.number,
   }
   constructor(props) {
     super(props);
@@ -33,7 +41,9 @@ export default class Textarea extends React.Component {
   }
   handleChange(e) {
     this.props.onChange(e);
-    this.resize();
+    if (!('value' in this.props)) {
+      this.resize();
+    }
   }
   render() {
     const { minRows, maxRows, ...otherProps } = this.props;